login and redoing some entities

+++ Atherapy/src/main/java/com/aTh/Atherapy/repository/MeetingRepo.java
@@ -3,8 +3,5 @@ package com.aTh.Atherapy.repository;
 import com.aTh.Atherapy.entity.Meeting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MeetingRepo extends JpaRepository<Meeting, Integer> {
-
-
-
+public interface MeetingRepo extends JpaRepository<Meeting, Long> {
 }
