+++ Atherapy/src/main/java/com/aTh/Atherapy/repository/RequestRepo.java
@@ -3,5 +3,5 @@ package com.aTh.Atherapy.repository;
 import com.aTh.Atherapy.entity.Request;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RequestRepo extends JpaRepository<Request, Integer> {
+public interface RequestRepo extends JpaRepository<Request, Long> {
 }
